+++ polyclonal/polyclonal.py
@@ -2466,6 +2466,172 @@ class Polyclonal:
             )
         return pd.DataFrame(result_files, columns=["epitope", "PDB file"])
 
+    def mut_icXX_df(
+        self,
+        *,
+        x,
+        icXX_col,
+        log_fold_change_icXX_col,
+        min_c=1e-5,
+        max_c=1e5,
+        logbase=2,
+        check_wt_icXX=(0.01, 100),
+        **kwargs,
+    ):
+        r"""Data frame of ICXX and log fold change induced by each mutation.
+
+        Parameters
+        ----------
+        x : float
+            Compute this ICXX, so 0.5 means IC50 aand 0.9 means IC90.
+        icXX_col : str
+            Name of column with ICXX values.
+        log_fold_change_icXX_col : str
+            Name of column with log fold change ICXX values.
+        min_c : float
+            Minimum allowed icXX, truncate values < this at this.
+        max_c : float
+            Maximum allowed icXX, truncate values > this at this.
+        logbase : float
+            Compute log fold-change in ICXX to this base.
+        check_wt_icXX : None or 2-tuple
+            If a 2-tuple, raise an error if the ICXX for the unmutated wildtype is
+            outside the range `(min_icXX, max_icXX)`. You may want to do this as
+            the clipping imposed by `min_c` and `max_c` will become a problem if
+            the wildtype ICXX is very different than one. In general, concentration
+            units should be used when fitting so that ICXX for wildtype is within an
+            order of magnitude (or at most two) from 1.
+
+        Returns
+        -------
+        pandas.DataFrame
+            Has columns "site", "wildtype", "mutant", `icXX_col`, and
+            "fold_change" + `icXX_col`. Note that the dataframe
+            **does** include entry for the wildtype (wildtype same as mutant)
+            at each site with a non-wildtype mutation.
+
+        """
+        # get ICXX for unmutated wildtype
+        wt_icXX = self.icXX(
+            pd.DataFrame({"aa_substitutions": [""]}),
+            x=x,
+            col="ICXX",
+            min_c=min_c,
+            max_c=max_c,
+        )["ICXX"].tolist()[0]
+        if check_wt_icXX:
+            if not (check_wt_icXX[0] <= wt_icXX <= check_wt_icXX[1]):
+                raise ValueError(f"{wt_icXX=} out of range {check_wt_icXX=}")
+
+        # get fold-change ICXX for all mutants
+        log_fold_change_icXX = (
+            self.icXX(
+                pd.DataFrame({"aa_substitutions": self.mutations),
+                x=x,
+                col="ICXX",
+                min_c=min_c,
+                max_c=max_c,
+            )
+            ).assign(
+                site=lambda x: x["mutation"].map(
+                    lambda m: self._mutparser.parse_mut(m)[1]
+                ),
+                mutant=lambda x: x["mutation"].map(
+                    lambda m: self._mutparser.parse_mut(m)[2]
+                ),
+                wildtype=lambda x: x["site"].map(self.wts),
+                log_fold_change_ICXX=lambda x: (
+                    numpy.log(x["ICXX"] / wt_icXX) / numpy.log(logbase)
+                )
+            )
+            [["site", "wildtype", "mutant", "ICXX", "log_fold_change_ICXX"]]
+        )
+        # add in the wildtypes all having a log fold change of zero
+        log_fold_change_icXX = pd.concat(
+            [
+                log_fold_change_icXX,
+                (
+                    log_fold_change_icXX
+                    [["site", "wildtype"]]
+                    .drop_duplicates()
+                    .assign(
+                        mutant=lambda x: x["wildtype"],
+                        ICXX=wt_icXX,
+                        log_fold_change_ICXX=0,
+                    )
+                ),
+            ],
+        )
+        assert (
+            len(log_fold_change_icXX)
+            == len(log_fold_change_icXX.groupby(["site", "mutant"]))
+        )
+        assert all(numpy.isfinite(log_fold_change_icXX["log_fold_change_icXX"]))
+
+        return (
+            log_fold_change_icXX
+            .rename(
+                columns={
+                    "ICXX": icXX_col, "log_fold_change_icXX": log_fold_change_icXX_col,
+                },
+            )
+            .sort_values(["site", "mutant"])
+            .reset_index(drop=True)
+        )
+
+    def mut_icXX_plot(
+        self,
+        *,
+        x=0.9,
+        icXX_col="IC90",
+        log_fold_change_icXX_col="log2_fold_change_IC90",
+        min_c=1e-5,
+        max_c=1e5,
+        logbase=2,
+        check_wt_icXX=(0.01, 100),
+        biochem_order_aas=True,
+        df_to_merge=None,
+        **kwargs,
+    ):
+        r"""Make plot of log fold-change in ICXX induced by each mutation.
+
+        Parameters
+        ----------
+        x : float
+            Same meaning as for :meth:`Polyclonal.mut_icXX_df`.
+        icXX_col : str
+            Same meaning as for :meth:`Polyclonal.mut_icXX_df`.
+        log_fold_change_icXX_col : str
+            Same meaning as for :meth:`Polyclonal.mut_icXX_df`.
+        min_c : float
+            Same meaning as for :meth:`Polyclonal.mut_icXX_df`.
+        max_c : float
+            Same meaning as for :meth:`Polyclonal.mut_icXX_df`.
+        logbase : floata
+            Same meaning as for :meth:`Polyclonal.mut_icXX_df`.
+        check_wt_icXX : None or 2-tuple
+            Same meaning as for :meth:`Polyclonal.mut_icXX_df`.
+        biochem_order_aas : bool
+            Biochemically order the amino-acid alphabet in :attr:`Polyclonal.alphabet`
+            by passing it through :func:`polyclonal.alphabets.biochem_order_aas`.
+        df_to_merge : None or pandas.DataFrame or list
+            To include additional properties, specify data frame or list of them which
+            are merged with :attr:`Polyclonal.mut_escape_df` before being passed
+            to :func:`polyclonal.plot.lineplot_and_heatmap`. Properties will
+            only be included in plot if relevant columns are passed to
+            :func:`polyclonal.plot.lineplot_and_heatmap` via `addtl_slider_stats`,
+            `addtl_tooltip_stats`, or `site_zoom_bar_color_col`.
+        **kwargs
+            Keyword args for :func:`polyclonal.plot.lineplot_and_heatmap`.
+
+        Returns
+        -------
+        altair.Chart
+            Interactive line plot and heatmap.
+
+        """
+        raise NotImplemented
+
     def mut_escape_plot(
         self,
         *,