+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 ---
 - Upgrade minimum ``altair`` to 5.0.0
 - Use ``ruff`` as linter rather than ``flake8``
+- Add ``Polyclonal.mut_icXX_df`` method
 
 5.3
 ---